requirements.txt
Django==3.1
Django-filter==2.4.0
djangorestframework==3.12.2
django-rest-auth==0.9.5
django-cors-headers==3.5.0
psycopg2-binary==2.8.6
gunicorn==20.0.4
whitenoise==5.2.0
python-decouple==3.4
virtualenv==20.4.2
yarn==0.0.9

